+++ spb/backends/__init__.py
@@ -1,7 +1,0 @@
-# TODO: while it would be nice to import the backends directly from
-# spb.backends, this would introduce a circular import. For the moment, let's
-# comment it out.
-# from spb.backends.matplotlib import MatplotlibBackend, MB
-# from spb.backends.plotly import PlotlyBackend, PB
-# from spb.backends.bokeh import BokehBackend, BB
-# from spb.backends.k3d import K3DBackend, KB